+++ cmake/modules/DuneGridMacros.cmake
@@ -23,8 +23,6 @@ include(AddAlbertaFlags)
 set_package_properties(Alberta PROPERTIES TYPE OPTIONAL
   PURPOSE "Provides the grid manager AlbertaGrid and file reader AlbertaReader")
 
-include(UseUG)
-
 set(DEFAULT_DGF_GRIDDIM 1)
 set(DEFAULT_DGF_WORLDDIM 1)
 set(DEFAULT_DGF_GRIDTYPE ONEDGRID)