+++ dune/grid/common/grid.hh
@@ -508,25 +508,11 @@ namespace Dune {
      */
     typedef typename GridFamily::Traits::LocalIdSet LocalIdSet;
 
-  protected:
-    template <class T>
-    using Communication_t = typename T::Communication;
-    template <class T>
-    using DeprecatedCollectiveCommunication_t = typename T::CollectiveCommunication;
-
-  public:
     /*! \brief A type that is a model of Dune::Communication.
        It provides a portable way for communication on the set
        of processes used by the grid.
      */
-    // if this line produces a warning then the Communication typedef is missing
-    // in the GridFamily::Traits
-    using Communication = detected_or_fallback_t< DeprecatedCollectiveCommunication_t,
-                                                  Communication_t, typename GridFamily::Traits>;
-
-    /** \deprecated Use Communication instead! Will be removed at some point in the future.
-     */
-    using CollectiveCommunication [[deprecated("CollectiveCommunication is deprecated, use Communication instead!")]] = Communication;
+    using Communication = typename GridFamily::Traits::Communication;
 
     //! Define type used for coordinates in grid module
     typedef ct ctype;
@@ -1073,10 +1059,6 @@ namespace Dune {
 
     /** \brief The type of the communication. */
     typedef CCType Communication;
-
-    /** \deprecated Use Communication instead! Will be removed at some point in the future. */
-    [[deprecated("Use Communication instead!")]]
-    typedef Communication CollectiveCommunication;
   };
 
   // Definition of capabilities for the interface class