+++ dune/grid/concepts/gridview.hh
@@ -93,7 +93,7 @@ requires(const GV gv, int codim, Dune::GeometryType type)
   { gv.indexSet()         } -> std::convertible_to<const typename GV::IndexSet&>;
   { gv.size(codim)        } -> std::convertible_to<int>;
   { gv.size(type)         } -> std::convertible_to<int>;
-  { gv.comm()             } -> std::convertible_to<typename GV::CollectiveCommunication>;
+  { gv.comm()             } -> std::convertible_to<typename GV::Communication>;
   { gv.overlapSize(codim) } -> std::convertible_to<int>;
   { gv.ghostSize(codim)   } -> std::convertible_to<int>;