+++ dune/grid/geometrygrid/gridfamily.hh
@@ -110,12 +110,6 @@ namespace Dune
 
         typedef typename HostGrid::Traits::Communication Communication;
 
-        /**
-         * \deprecated Use Communication instead! Will be removed after Dune 2.9.
-         */
-        [[deprecated("Use Communication instead!")]]
-        typedef Communication CollectiveCommunication;
-
         typedef Dune::GridView< GeoGrid::GridViewTraits< typename HostGrid::LeafGridView, CoordFunction, Allocator > > LeafGridView;
         typedef Dune::GridView< GeoGrid::GridViewTraits< typename HostGrid::LevelGridView, CoordFunction, Allocator > > LevelGridView;
       };