+++ dune/grid/geometrygrid/gridview.hh
@@ -59,12 +59,6 @@ namespace Dune
 
       typedef typename HostGridView::Communication Communication;
 
-      /**
-       * \deprecated Use Communication instead! Will be removed after Dune 2.9.
-       */
-      [[deprecated("Use Communication instead!!")]]
-      typedef Communication CollectiveCommunication;
-
       template< int codim >
       struct Codim
       {
@@ -111,12 +105,6 @@ namespace Dune
       typedef typename Traits::IntersectionIterator IntersectionIterator;
 
       typedef typename Traits::Communication Communication;
-
-      /**
-       * \deprecated Use Communication instead! Will be removed after Dune 2.9.
-       */
-      [[deprecated("Use Communication instead!!")]]
-      typedef Communication CollectiveCommunication;
 
       template< int codim >
       struct Codim