+++ dune/grid/onedgrid/onedgridviews.hh
@@ -50,10 +50,6 @@ namespace Dune
     /** \brief type of the communication */
     typedef typename Grid :: Traits :: Communication Communication;
 
-    /** \deprecated Use Communication instead! Will be removed after Dune 2.9. */
-    [[deprecated("Use Communication instead!")]]
-    typedef Communication CollectiveCommunication;
-
     template< int cd >
     struct Codim
     {
@@ -102,10 +98,6 @@ namespace Dune
 
     /** \brief type of the communication */
     typedef typename Traits :: Communication Communication;
-
-    /** \deprecated Use Communication instead! Will be removed after Dune 2.9. */
-    [[deprecated("Use Communication instead!")]]
-    typedef Communication CollectiveCommunication;
 
     /** \brief Codim Structure */
     template< int cd >
@@ -240,10 +232,6 @@ namespace Dune
     /** \brief type of the communication */
     typedef typename Grid :: Traits :: Communication Communication;
 
-    /** \deprecated Use Communication instead! Will be removed after Dune 2.9. */
-    [[deprecated("Use Communication instead!")]]
-    typedef Communication CollectiveCommunication;
-
     template< int cd >
     struct Codim
     {
@@ -291,10 +279,6 @@ namespace Dune
 
     /** \brief type of the communication */
     typedef typename Traits :: Communication Communication;
-
-    /** \deprecated Use Communication instead! Will be removed after Dune 2.9. */
-    [[deprecated("Use Communication instead!")]]
-    typedef Communication CollectiveCommunication;
 
     /** \brief Codim Structure */
     template< int cd >